VPN-3508 - part 13 - unit-test app
