+++ CMakeLists.txt
@@ -151,6 +151,7 @@ if(NOT CMAKE_CROSSCOMPILING)
     add_subdirectory(tests/nativemessaging EXCLUDE_FROM_ALL)
     add_subdirectory(tests/unit EXCLUDE_FROM_ALL)
     add_subdirectory(tests/qml EXCLUDE_FROM_ALL)
+    add_subdirectory(src/apps/unit_tests EXCLUDE_FROM_ALL)
 
     # Dummy Testing Client
     add_subdirectory(tests/dummyvpn EXCLUDE_FROM_ALL)
