+++ src/apps/unit_tests/appconstants.h
@@ -12,11 +12,13 @@ namespace AppConstants {
 // This is used by SettingsHolder to configure the QSetting file.
 constexpr const char* SETTINGS_APP_NAME = "FooBar";
 
+// The localization filename prefix. The real file name should be called:
+// `LOCALIZER_FILENAME_PREFIX` + '_' + languageCode + ".qm". For instance:
+// `mozillavpn_it.qm
+constexpr const char* LOCALIZER_FILENAME_PREFIX = "foo_bar";
+
 // The prefix for the user-agent requests
 constexpr const char* NETWORK_USERAGENT_PREFIX = "FooBar";
-
-// The file name for the logging
-constexpr const char* LOG_FILE_NAME = "foobar.txt";
 
 #if defined(__APPLE__)
 // This is the name of the service to encrypt the settings file
@@ -37,12 +39,10 @@ constexpr const wchar_t* WINDOWS_CRED_KEY = L"Foo bar";
 constexpr const char* ANDROID_LOG_NAME = "foobar";
 #endif
 
-#if defined(MZ_LINUX)
 constexpr const char* LINUX_CRYPTO_SETTINGS_KEY =
     "org.mozilla.foobar.cryptosettings";
 constexpr const char* LINUX_CRYPTO_SETTINGS_DESC =
     "FooBar settings encryption key";
-#endif
 
 // Communication pipe between instances
 #if defined(MZ_WINDOWS)
