+++ src/apps/unit_tests/appfeaturelistcallback.h
@@ -1,0 +1,5 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+// NOTE! Do not include this file directly. Use featurelistcallback.h instead.
