+++ src/apps/unit_tests/helper.h
@@ -1,0 +1,24 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#ifndef HELPER_H
+#define HELPER_H
+
+#include <QObject>
+#include <QVector>
+#include <QtTest/QtTest>
+
+class TestHelper : public QObject {
+  Q_OBJECT
+
+ public:
+  TestHelper();
+
+ public:
+  static QVector<QObject*> testList;
+
+  static QObject* findTest(const QString& name);
+};
+
+#endif  // HELPER_H
