+++ src/apps/unit_tests/themes/themes.qrc
