+++ src/shared/resources/license.qrc
@@ -1,5 +1,5 @@
 <RCC>
     <qresource prefix="/license">
-        <file alias="LICENSE.md">../../../../LICENSE.md</file>
+        <file alias="LICENSE.md">../../../LICENSE.md</file>
     </qresource>
 </RCC>
