+++ src/shared/settingslist.h
@@ -29,6 +29,116 @@
 #define SETTING_STRINGLIST(getter, ...) \
   SETTING(QStringList, toStringList, getter, __VA_ARGS__)
 
+SETTING_BOOL(addonCustomServer,        // getter
+             setAddonCustomServer,     // setter
+             removeAddonCustomServer,  // remover
+             hasAddonCustomServer,     // has
+             "addon/customServer",     // key
+             false,                    // default value
+             false,                    // user setting
+             false                     // remove when reset
+)
+
+SETTING_STRING(addonCustomServerAddress,        // getter
+               setAddonCustomServerAddress,     // setter
+               removeAddonCustomServerAddress,  // remover
+               hasAddonCustomServerAddress,     // has
+               "addon/customServerAddress",     // key
+               AppConstants::addonBaseUrl(),    // default value
+               false,                           // user setting
+               false                            // remove when reset
+)
+
+SETTING_BOOL(addonProdKeyInStaging,        // getter
+             setAddonProdKeyInStaging,     // setter
+             removeAddonProdKeyInStaging,  // remover
+             hasAddonProdKeyInStaging,     // has
+             "addon/prodKeyInStaging",     // key
+             false,                        // default value
+             false,                        // user setting
+             false                         // remove when reset
+)
+
+SETTING_STRINGLIST(featuresFlippedOff,        // getter
+                   setFeaturesFlippedOff,     // setter
+                   removeFeaturesFlippedOff,  // remover
+                   hasFeaturesFlippedOff,     // has
+                   "featuresFlippedOff",      // key
+                   QStringList(),             // default value
+                   false,                     // user setting
+                   false                      // remove when reset
+)
+
+SETTING_STRINGLIST(featuresFlippedOn,        // getter
+                   setFeaturesFlippedOn,     // setter
+                   removeFeaturesFlippedOn,  // remover
+                   hasFeaturesFlippedOn,     // has
+                   "featuresFlippedOn",      // key
+                   QStringList(),            // default value
+                   false,                    // user setting
+                   false                     // remove when reset
+)
+
+SETTING_DATETIME(installationTime,        // getter
+                 setInstallationTime,     // setter
+                 removeInstallationTime,  // remover
+                 hasInstallationTime,     // has
+                 "installationTime",      // key
+                 QDateTime(),             // default value
+                 false,                   // user setting
+                 false                    // remove when reset
+)
+
+SETTING_STRING(installedVersion,        // getter
+               setInstalledVersion,     // setter
+               removeInstalledVersion,  // remover
+               hasInstalledVersion,     // has
+               "installedVersion",      // key
+               "",                      // default value
+               false,                   // user setting
+               false                    // remove when reset
+)
+
+SETTING_STRING(languageCode,        // getter
+               setLanguageCode,     // setter
+               removeLanguageCode,  // remover
+               hasLanguageCode,     // has
+               "languageCode",      // key
+               "",                  // default value
+               true,                // user setting
+               false                // remove when reset
+)
+
+SETTING_STRING(previousLanguageCode,        // getter
+               setPreviousLanguageCode,     // setter
+               removePreviousLanguageCode,  // remover
+               hasPreviousLanguageCode,     // has
+               "previousLanguageCode",      // key
+               "",                          // default value
+               true,                        // user setting
+               false                        // remove when reset
+)
+
+SETTING_BOOL(systemLanguageCodeMigrated,        // getter
+             setSystemLanguageCodeMigrated,     // setter
+             removeSystemLanguageCodeMigrated,  // remover
+             hasSystemLanguageCodeMigrated,     // has
+             "systemLanguageCodeMigrated",      // key
+             false,                             // default value
+             false,                             // user setting
+             true                               // remove when reset
+)
+
+SETTING_DATETIME(updateTime,        // getter
+                 setUpdateTime,     // setter
+                 removeUpdateTime,  // remover
+                 hasUpdateTime,     // has
+                 "updateTime",      // key
+                 QDateTime(),       // default value
+                 false,             // user setting
+                 false              // remove when reset
+)
+
 SETTING_STRING(userEmail,        // getter
                setUserEmail,     // setter
                removeUserEmail,  // remover
@@ -37,6 +147,18 @@ SETTING_STRING(userEmail,        // getter
                "",               // default value
                false,            // user setting
                true              // remove when reset
+)
+
+#define DEFAULT_THEME "main"
+
+SETTING_STRING(theme,          // getter
+               setTheme,       // setter
+               removeTheme,    // remover
+               hasTheme,       // has
+               "theme",        // key
+               DEFAULT_THEME,  // default value
+               true,           // user setting
+               true            // remove when reset
 )
 
 // The app must implement its settings list file.
